This commit was generated by cvs2svn to compensate for changes in r11917, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@11918 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/plus4/ted-badline.c
@@ -43,9 +43,9 @@ void ted_badline_check_state(BYTE value, int cycle, unsigned int line)
 
     /* Check whether bad line state has changed.  */
     was_bad_line = (ted.allow_bad_lines
-                    && (ted.raster.ysmooth == (line & 7)));
+                    && (ted.raster.ysmooth == (int)(line & 7)));
     now_bad_line = (ted.allow_bad_lines
-                    && ((value & 7) == (line & 7)));
+                    && ((int)(value & 7) == (int)(line & 7)));
 
     if (was_bad_line && !now_bad_line) {
